Remove Instacart
I applied to Instacart through this list, but won't be going forward with their process. The recruiter, documents that recruiter sent, and further research on Blind, Glassdoor, etc. indicate that they do tech screens through a 3rd party (karat), so you don't even talk to an engineer at the company. Apparently in that screen CS knowledge/topic questions are asked and you're expected to get through 2 leetcode-esque questions in 30 or so minutes. None of this "resembles day-to-day work" to me. 

I don't think that this process fits the spirit of this list at all, but at the very least there's a mismatch in the real process and what's describes here.